* ext/bigdecimal/bigdecimal.c (BigDecimal_hash): st_index_t may not be   fixable on 64bit mswin/mingw.
* ext/date/date_core.c (d_lite_hash): ditto.
  [Backport #13877]

* ext/openssl/ossl_bn.c (ossl_bn_hash): ditto.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_2_3@59879 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/date/date_core.c
@@ -6405,7 +6405,7 @@ d_lite_hash(VALUE self)
     h[2] = m_df(dat);
     h[3] = m_sf(dat);
     v = rb_memhash(h, sizeof(h));
-    return LONG2FIX(v);
+    return LONG2FIX((long)v);
 }
 
 #include "date_tmx.h"
